minor bug fixes. rev, get. best version

+++ main.h
@@ -21,8 +21,8 @@
 int init_all(mpu9250*);
 void init_clock();
 void stabilize_imu(mpu9250*);
-int get_pitch(mpu9250*);
-void my_itoa(int);
+float get_pitch(mpu9250*);
+void my_itoa(float);
 void pid(float);
 
 const Timer_A_ContinuousModeConfig continuousModeConfig =
